# docs/20260424-openapi-swagger.yaml
openapi: 3.1.0
info:
  title: Health Screening Interpreter Chatbot
  description: >
    An intelligent chatbot API to help understand health screening results.
    Uses Retrieval-Augmented Generation (RAG) with vector-based document retrieval
    and LLM-powered answer generation.
  version: 0.1.0
  contact:
    name: API Support
servers:
  - url: http://localhost:8000
    description: Local development
security:
  - ApiKeyAuth: []
  - {}
...
paths:
...
  /documents:
    get:
      operationId: listDocuments
      tags: [Documents]
      summary: List indexed documents
      description: Get a paginated summary of all indexed document chunks from ChromaDB.
      parameters:
        - name: limit
          in: query
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 500
            default: 50
          description: Maximum number of items to return
        - name: offset
          in: query
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
          description: Number of items to skip
        - name: source_type
          in: query
          required: false
          schema:
            type: string
          description: Filter by source type (e.g., pdf, html)
      responses:
        "200":
          description: Paginated list of document chunks
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentListResponse"
        "503":
          description: Vector store not initialized
  /documents/{doc_id}:
    get:
      operationId: getDocument
      tags: [Documents]
      summary: Get a single document chunk
      description: Retrieve full content and metadata for a specific chunk.
      parameters:
        - name: doc_id
          in: path
          required: true
          schema:
            type: string
          description: Document chunk ID
      responses:
        "200":
          description: Document chunk details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DocumentDetailResponse"
        "404":
          description: Document not found
        "503":
          description: Vector store not initialized
...
  /experiments/{experiment_name}/results:
    get:
      operationId: getExperimentResults
      tags: [Experiments]
      summary: Get experiment results
      description: Get the full JSON report for a specific experiment.
      parameters:
        - name: experiment_name
          in: path
          required: true
          schema:
            type: string
          description: Experiment name (matches YAML config filename stem)
      responses:
        "200":
          description: Experiment results report
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "404":
          description: Report not found
  /experiments/{experiment_name}/config:
    get:
      operationId: getExperimentConfig
      tags: [Experiments]
      summary: Get parsed experiment config
      description: Get the parsed YAML configuration for a specific experiment.
      parameters:
        - name: experiment_name
          in: path
          required: true
          schema:
            type: string
          description: Experiment name (matches YAML config filename stem)
      responses:
        "200":
          description: Parsed experiment config
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        "404":
          description: Experiment config not found
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >
        API key for authentication. Required on all endpoints except `/`, `/health`, `/docs`, `/openapi.json`.
        Keys are configured via `API_KEYS_JSON` or `API_KEYS` environment variables.
  schemas:
...
    DocumentListItem:
      type: object
      properties:
        id:
          type: string
        source:
          type: string
        page:
          type: integer
          nullable: true
        source_type:
          type: string
        source_class:
          type: string
        content_type:
          type: string
        content_preview:
          type: string
          description: First 200 characters of the chunk content
        content_length:
          type: integer
    DocumentListResponse:
      type: object
      properties:
        total:
          type: integer
        offset:
          type: integer
        limit:
          type: integer
        items:
          type: array
          items:
            $ref: "#/components/schemas/DocumentListItem"
        source_type_counts:
          type: object
          additionalProperties:
            type: integer
        index_metadata:
          type: object
          additionalProperties: true
    DocumentDetailResponse:
      type: object
      properties:
        id:
          type: string
        content:
          type: string
        metadata:
          type: object
          additionalProperties: true
        content_length:
          type: integer
...